Corrected genotyping QC list code for non-existent design
A test was failing in t/70 following the outer join SQL amendment to
bring back all rows in a plate even if there is no genotyping QC data
associated. This was owing to ancestral wells not being available in
test fixture data.

The code has been corrected in genotyping QC so that shuld thi shappen
in a live situation '-' symbols will be displayed under the design
heading and no gene symbol will be available.