Added distribute and override columns to genotyping_qc view grid
Refactored browse code. All methods for dealing with the model side of
the grid are now in the GenotypingQC plugin. There is only one method to
update in one place if further columns need to be added in future.

Added the ability to reset attribute values across the grid, where
appropriate, otherwise priority ranking is applied.