Added automated re-tries when a seq or pcr primer search fails
There are two situations where a retry (fresh attempt) is required:
1) for seq primers and pcr primers, where primer3 found no primers.
Here the search window is moved along and the dead field is incremented.
The job will be retried up to four times. Substantial logging added to
show what is going on.

2) for pcr primers, where primer3 found primers but the genomic check
returned non unique alignments. The effect is the same, the windows are
slid along as in (1) but, the cause is different - so this is noted in
the log.

The starting parameters for search and dead fields have been set to
correspond with the static values in the previous versions, so the first
attempt primers should be the same as would have been generated
previously. However, any minor variation in parameters can have a
profound effect upon primer generation, so identity cannot be
guaranteed.