Corrected the reporting of final_pick_well_accepted in gene summary reporting
The gene summary report was not showing well_accepted according to the $well->is_accepted
method even though this had been correctly picked upi in the summaries
reporting. This is because the woring summary attribute was being used
(final_well_accepted was being used instead of
final_pick_well_accepted).