Public sponsor sub-reports should not disclose 'info' comments
'info' comments for public sponspor sub-reports should only be included
in the csv download for logged-in users. The controller has been updated
to include a test for logged in user when the csv file is downloaded.
If the user is not logged in the info comments field is filtered out.

Perhaps, this should not happen in the controller. However, caching the
logged out csv file separately is harder to achieve. And we are not
interacting with the model at this point, so a util method seems
overkill.