Changed code for bringing back most recent Miseq projects to use DBIx::Class order by
It's best practice to delegate sorting and searching to the underlying
database. This commit uses SQL abstract to request the database to
return the 20 most recent projects order by date. The original code
could not guarantee that this would happen as there was no order by on
the search statement. We also avoid havig to use Perl for the sort step.