get_well_exp_graphs controller fixed
Updated the user pointmutation controller to be able to get the well
experiment graphs using only the database instead of the file system.
The old method is still there, right above the new one (_old). More data
can be passed through the array of experiments, and I would have to
evaluate how helpful that would be. At the moment it is only passing
what the old method was passing, but it will need to be revisited in the
future.
Note: Uncommented the following line:
            delete $exp_params->{well_id};
It is on line 308 of the User/PointMutation file. It was previously
commented out for unknown reasons.