Commit before db replication
Wrong data inserted. Count values from the quantfication files were
inserted to the miseq experiment table, rather than the
miseq_well_experiment table. That means that it now necessary to
replicate local db and rework the issue. The steps that need to be taken
now to fix the issue are the following:
-Add nhej, total, hdr and wt counts on the miseq well exp rather than
miseq exp. That means going inside the ddl files.
-Recreate the audit file
-Schema needs to be regenerated and the methds within the schema files
as well. Should probably revert the changes already made back to the
live version and work it from there
-Change the migration scripts. They are located inside bin at the
moment. They are utilizing a library located in Model::Util called
ImportCrispressonQC. Functionality needs to be changed in order to
update the correct entities.
-Rework the controller for the targeting efficiency