Reverted drop gene column on miseq_experiment
Reverted due to lack of experiment_ids in the table. These will be
updated through out the backfill process however we do not have all the
parent plates in LIMS2 as of yet. Without the gene nor the experiment,
the lab would lose an important feature on the miseq interface until the
backfill is complete. At the current moment, this is an unknown length
of time. Once the backfill is complete, we'll look back into dropping
the gene column