Changes in the way data is handled
Due to having to work on a different project I had to commit here. I
changed the way data is being stored for the summary data ref. It is
better but it still isnt exactly what the json decoder expects. Ill have
to check the live version and then get back to this and fix it.