Added Phred33 scoring to Miseq Alleles page
Removed the simple base highlighting mode and replaced it with Phred33 quality
highlighting. Scores are an average of how the base performed across all
of the reads. If the base is below either score threshold, it'll be
highlighted to the user.
Also cleaned up the page and included a new colour key which will update
depending on which mode the use is in.