Initial amplicon table design
Extracted design_id away from the main table. This has been placed in a
linker table. Reason for this is to prevent a headache down the line
when we begin work on the external plates. These will need to reuse
these tables but they will not have a design ID.