Fix dependency issue with test fixtures.
A new foreign key reference from cell-line to species had been added.
This meant that the test setup for the species table had to be run
before that for cell-line.

Note that after applying this change some tests are being flagged up
as failing that weren't previously:

  - 10_LIS2_Model_Util_Trivial
  - 10_LIMS2_Report_GeneEPSummary
  - 10_LIMS2_SummaryGeneration_SummariesWellLocal
  - 10_LIMS_WebApp_Controller_Auth

However, after making the change more tests are also being run, so
I'm assuming it doesn't actually cause any tests to break.