Runs tests on each push to gitlab.
This basically runs prove -lr t in a pre-built environment. Although
this is a first good step, it is of somewhat limited utility as the
environment in which it runs is highly divergent from the production
environment.

The scripts should be run as ubuntu user. This requires that both the
gitlab-runner service runs as ubuntu, and that this service runs the
script as ubuntu.

The former is done by adding
User=ubuntu
Group=ubuntu
under the [Service] section of
/etc/systemd/system/gitlab-runner.service

and the latter involves adding
ExecStart=/usr/bin/gitlab-runner "run" "--working-directory" "/home/ubuntu" "--config" "/etc/gitlab-runner/config.toml" "--service" "gitlab-runner" "--user" "ubuntu"