Install perl modules with broken tests.
Some of the perl modules were failing to install, due to the tests failing.
Looking at the failures it appears that the problems were with the test
code, rather than the code under test, so I have installed these modules
without tests. Both modules have fixes in the pipeline, so we should be
able to revert this commit soon.