Create distinct web-app and test docker images.
This is enough code to get the tests running (but not passing) in
a docker container.

The desired endpoint is that the web-app and test images should be
identical, expect for the things for which they shouldn't - command
run, database connected to, perhaps some other things.

However, given that the high-level image has been running in production
without major problems, my intermediate strategy is to keep the web-app
image the same as this for the time being, and then bring the two images
(web-app and test) in to alignment when all the tests are passing. This
will allow me to do it in a series of small steps.

See also the related PRs:
https://gitlab.internal.sanger.ac.uk/sci/openstackscripts/-/merge_requests/new?merge_request%5Bsource_branch%5D=get-tests-running
lims2-webapp-filesystem!11 (merged)

Extra notes: