Use cache parameter in public report tests.
As it was, the test were failing unless there was a cached file in the
right location. It was possible to ensure that this file was there for
the purpose of testing, but it seemed more in the spirit of the test
to run the controller function without caching. This is what happens,
for comparison, when one clicks the 'Switch to Mouse' button in the
user interface.