Use a non-existent file for rest config.
I set the LIMS2_REST_CLIENT_CONFIG to a non-existent file. This is
to allow the tests to run, without having access to the production
system from a dev environment - which could cause problems.

What the *actual* config should look like for a dev/test environment
is something to be considered later.