Check for missing loci data.
When the primer design code is unable to find oligo sequences in the
expected locations we get an `undef` being returned by `loci_builder`.
This means that at the point of being validated we have something like:
```
  'exf' => {
	     'loci' => {
			 'chr_strand' => 1,
			 'assembly' => 'GRCh38'
		       },
	     'seq' => 'AGACATCAAGAGGAGAACTGC'
	   },
```
instead of like
```
  'exf' => {
	     'loci' => {
			 'assembly' => 'GRCh38',
			 'chr_start' => 46066725,
			 'chr_end' => 46066745,
			 'chr_strand' => 1,
			 'chr_name' => '17'
		       },
	     'seq' => 'TCCCCTTGTATAGTCTTCAGC'
	   },
```
which resulted in an erroneous 'Oligos have inconsistent chromosomes'
error message in the _validate_primers function - because
$primers->{'exf'}{l'oci'}{'chr_name'} is undef.

This commit adds extra validation to test that the expected loci data is
present.

Ideally, we would bbe able to report to the user that the software was
unable to locate the oligo sequence, however it's difficult to propogate
this information from where the error happens to where it is reported,
so we just report that data is missing.