Set ensembl host to standard EBI url.
The interface to the ensembl API is instantiated in at least two places.
The first one falls back to ensembl's own site if the registry fails to
load for any reason, whereas the second one only falls back if the
env-vars aren't set. Setting the env-var to the standard url means that
both _should_ use the same url.