Use plate->species_id to get species information.
This gets the test to pass, but I'm very suspicious that this isn't the
right way to get species information - the plate table just doesn't seem
to be the right place to store this. Perhaps the plate is always created
with species_id = "Human" just because this is a non-nullable field?