Refactor QUERY_INHERITED_EXPERIMENT query.
This makes the query easier to understand (I hope).

The hardest part to understand is the recursion part, which constructs
the backward component of the well-process graph for wells in a particualar
plate.

This separates the calculation of this graph from the extraction of
experiment, design, and crispr data.

Note that this changes some of the column names:
	wh.crispr_id -> pc.crispr_id
	wh.design_id -> pd.design_id
but these column neames are not used by the consuming code.