Add gene ID to general-info.
Abstractly it seems that more than one gene can be associated to a well.
I'm still awaiting confirmation from the lab as to whether this happens
in practice. However, it makes sense to just limit the function to only
deal with situtations where one gene is associted to the well, for the
time being.