Use stringification of $well to provide clone ID.
The Well class overides the stringify operator to return the clone ID,
i.e. a string of the form <plate_name>_<well_name>, (at least in case
where the plate exists.)

This is a bit more 'canonical' than just using an ad-hoc method to
create the ID, as in the previous commit, but it is also a bit more
cryptic.