Fix bug when getting gene count.
Attempting to spin up a docker container was causing the following
error:

ERROR Catalyst - Attribute (sponsor_gene_count) does not pass the type
constraint because: Validation failed for 'ArrayRef' with value undef at
reader LIMS2::Model::Util::SponsorReportII::sponsor_gene_count

Investigating, this was due to me running the container using a database
with a more minimal set of data than usual (for development and testing
purposes). In this case, the `_build_sponsor_gene_count` function
returns an `undef` rather than an array ref, as expected, as there is no
data with which to populate the array. This fixes things by returning an
empty array ref when the data does not exist.