Use commit ID for test image tags in CI.
Previously we were just using the tag 'test' when building test images.
As the build and test stages ran as separate jobs, there was a chance
that we might have:

	build branch A
	build branch B
	test branch A
	test branch B

In such a case the 'test branch A' bit would *actually* be testing using
the branch B image. Changing to tagging using commit ID means that this
is no longer the case.