Remove reference to test tag in CI script.
When doing the work in 9a3881de473b I failed to remove these two lines,
which referenced a no longer used 'test' tagged image. Although these
lines didn't do anything useful, the script continued to work as long as
the test-tagged image was in the registry. The image has now been
auto-removed from the registry, and the CI script was failing. Hence
this needs to be removed.