Allow clicking on quadrant text on point mutation page.
Previously, the overlay text for the quadrants was capturing click
events, which meant that clicking on the text did not take one to the
overview for that plate. As the text was at the centre of the quadrant -
the most likely place one would click - this was problematic. One can
now click anywhere in the quadrant to view the overview.