Move fixtures from CrisprEsQcRun to CrisprEsQcRun.
Previously, data was dumped to CrisprEsQcRun.csv, but loaded from
CrisprEsQcRuns.csv. This has been fixed in changes to LIMS-2 utils, so
that data is now dumped to CrisprEsQcRuns.csv, which is in-line with the
module name. This moves the data that was previusly created by the
dumping process in to the right place. It over-writes an older file that
was in that place.