Add zero indel to indel data
The zero-indel, i.e. the indel with zero insertions / deletions /
mutations, is not include in the data in the IndelHistogram table.
Instead it has to be computed by removing the sum of the frequencies of
the non-zero indels and subtracting this from the total reads.

Note that the code in this commit replicates a lot of the code in
lib/LIMS2/WebApp/Controller/User/PointMutation.pm. However, it seems
that the page that uses that code is the one that Katie is using to
test that what I'm doing here is correct, so extracting the common code
is problematic.