Add shim for indel data input
The buildIndelPlot function assumes that the data exists for all indels
within a certain range, even if the frequency is zero. This pre-process
the data to put it in that form.

Ideally, this should be called internally by buildIndelPlot. However,
because this piece of work is being tested by comparing the plot on the
point-mutation page, I don't want to do anything that might possibly
change that for the time-being.

I believe however, that this function *can* be called internally by
buildIndelPlot without changing its behaviour. So I will probably move
it later.