Install newest version of postgres client.
Following on from the recent postgres upgrade
https://jira.sanger.ac.uk/browse/LIMS-89
it was necessary to upgrade the client, to allow dumping amonst other
things.

I did this copying the instructions from
https://www.postgresql.org/download/linux/ubuntu/

It would seem more logical for this part to be added in the dockerfile
after the other os-level dependencies, but before the perl ones. However
this would invalidate dockers build caching and cause the perl
dependencies to be added anew, which could have unintended consequences,
so I've put it at the bottom of the file.