Add initial missing miseq table.
This is a table of all the clones on any FP plate for which:

  * All the clones on that plate have a single mapping between
    FP and PIQ plate,
  * All the clones on that plate have no mapping from the PIQ plate
    to a MISEQ plate.

The idea is that the missing miseq data will be filled in by those in
the lab, and then we will use this table to back-fill all the missing
data in LIMS2.

We've chosen this particular subset of clones as a first step as this
case is seen as the most tractable.