Use a single `get_clones` function.
Previously, there was a `get_clones` in both the check_it module, and
the main script module. They returned the data in slightly different
form, but more importantly, the one used in `check_it` returned a list
with repreats. This has been fixed and now each element of the returned
collection is unique.