Add endpoint to get out experiment id
Note that the logic for obtaining the experiment from the clone ID is
quite involved, so it made sense to to use the pre-existing perl code to
avoid subtle differences.

This is only needed as part of the work to fix up the clone data. See
comment in code for more details.